Round out game logic
- Implement functions & event handlers that render
  or remove player options
- When user quits the ui will be cleared of all
  results/options and the score will be 0 - 0
- When user starts new game options will appear

This works for a game where first to 5 points wins
but there are some issues.

1. Once the game is over, when new game button is
pressed the UI is reset. However, clicking rock,
paper or scissors does not trigger anything since
the event handlers have been removed. The options
only work when pressing quit before pressing new
game which is not what we want.

2. The playRound function poses memory issues
since it is recreating the same object every time
it is called.

These issues will be fixed. This is just a point
where everything is close enough to create of a
snapshot.